compiler,runtime: do more direct interfaces
A direct interface is an interface whose data word contains the
actual data value, instead of a pointer to it. The gc toolchain
creates a direct interface if the value is pointer shaped, that
includes pointers (including unsafe.Pointer), functions, channels,
maps, and structs and arrays containing a single pointer-shaped
field. In gccgo, we only do this for pointers. This CL unifies
direct interface types with gc. This reduces allocations when
converting such types to interfaces.

Our method functions used to always take pointer receivers, to
make interface calls easy. Now for direct interface types, their
value methods will take value receivers. For a pointer to those
types, when converted to interface, the interface data contains
the pointer. For that interface to call a value method, it will
need a wrapper method that dereference the pointer and invokes
the value method. The wrapper method, instead of the actual one,
is put into the itable of the pointer type.

In the runtime, adjust funcPC for the new layout of interfaces of
functions.

Change-Id: Ib4869d0140156c2a90d3af6eff7ce74f0f3bc7ac
Reviewed-on: https://go-review.googlesource.com/c/gofrontend/+/168409
Reviewed-by: Ian Lance Taylor <iant@golang.org>