runtime: persistentalloc and cache itabs
Previously, each time we do an interface conversion for which the
method table is not known at compile time, we allocate a new
method table.

This CL ports the mechanism of itab caching from the gc runtime,
adapted to our itab representation and method finding mechanism.
With the cache, we reuse the same itab for the same (interface,
concrete) type pair. This reduces allocations in interface
conversions.

Unlike the gc runtime, we don't prepopulate the cache with
statically allocated itabs, as currently we don't have a way to
find them. This means we don't deduplicate run-time allocated
itabs with compile-time allocated ones. But that is not too bad
-- it is just a cache anyway.

As now itabs are never freed, it is also possible to drop the
write barrier for writing the first word of an interface header.
I'll leave this optimization for the future.

Change-Id: I4a0e85efe7afad575ed4ed60fdb24b9bfdb65fec
Reviewed-on: https://go-review.googlesource.com/c/gofrontend/+/171617
Reviewed-by: Ian Lance Taylor <iant@golang.org>