libgo/testsuite: ignore symbols with a leading dot in symtogo
On AIX, a function has two symbols, a text symbol (with a leading dot)
and a data one (without it).
As the tests must be run only once, only the data symbol can be used to
retrieve the final go symbol. Therefore, all symbols beginning with a dot
are ignored by symtogo.

Change-Id: Ic3b3ef5d9bb98e6848bcf6526e8724bdbc83b37a
Reviewed-on: https://go-review.googlesource.com/c/gofrontend/+/177837
Reviewed-by: Ian Lance Taylor <iant@golang.org>