compiler: make use of specialized fast map routines
In the runtime there are specialized fast map routines for
certain kep types. This CL lets the compiler make use of these
functions, instead of always using the generic ones.

As we now generate multiple versions of map delete calls, to make
things easier we delay the expansion of the built-in delete
function to flatten phase.

Change-Id: Ie246a8fdbe9a69816733bae6ca9e14af6634b65a
Reviewed-on: https://go-review.googlesource.com/c/gofrontend/+/180858
Reviewed-by: Ian Lance Taylor <iant@golang.org>