runtime: scan write barrier buffer conservatively
In gccgo, we insert the write barriers in the frontend, and so we
cannot completely prevent write barriers on stack writes. So it
is possible for a bad pointer appearing in the write barrier
buffer. When flushing the write barrier, treat it the same as
sacnning the stack. In particular, don't mark a pointer if it
does not point to an allocated object. We already have similar
logic in greyobject. With this, hopefully, we can prevent an
unallocated object from being marked completely.

Change-Id: If8349916b648b1b4241b8499f57c764540402bdf
Reviewed-on: https://go-review.googlesource.com/c/gofrontend/+/190599
Reviewed-by: Ian Lance Taylor <iant@golang.org>