compiler, runtime: support and use single argument go:linkname
The gc compiler has started permitting go:linkname comments with a
single argument to mean that a function should be externally visible
outside the package.  Implement this in the Go frontend.

Change the libgo runtime package to use it, rather than repeating the
name just to export a function.

Remove a couple of unnecessary go:linkname comments on declarations.

Change-Id: Ie0feafd0be8498f7eab352dab09318ba4689f808
Reviewed-on: https://go-review.googlesource.com/c/gofrontend/+/192197
Reviewed-by: Cherry Zhang <cherryyz@google.com>
Reviewed-by: Than McIntosh <thanm@google.com>