compiler: don't report runtime escapes if we've seen errors
If we get errors during compilation, we skip the escape analysis pass.
If we are compiling the runtime package, we report an error if a bound
method expression escapes.  The effect is that if we get an error
while compiling the runtime package, we would report confusing and
meaningless errors about bound method expressions escaping.

This CL stops doing that.

Change-Id: I3d5f413c6fe2c48f90439f845e7be18db23e124e
Reviewed-on: https://go-review.googlesource.com/c/gofrontend/+/192720
Reviewed-by: Cherry Zhang <cherryyz@google.com>