libgo: support gollvm build on arm64 linux
This CL serves as part of an initial change for enabling gollvm
building on arm64 linux, the rest of the change will be covered by
another one to the gollvm repo.

Incorporate type definition of 'uint128' to 'runtime' and 'syscall'
packges, the change is not specific to arm64 linux but made available
for all platforms.

Verified by building and unit-testing gollvm on linux x86-64 and arm64.

Verified by building and checking gccgo on linux x86-64 and arm64.

Fixes golang/go#33711

Change-Id: I4720c7d810cfd4ef720962fb4104c5641b2459c0
Reviewed-on: https://go-review.googlesource.com/c/gofrontend/+/194997
Reviewed-by: Ian Lance Taylor <iant@golang.org>