compiler: don't use predeclared location for implicit runtime import
For the main package we add an implicit import of the runtime package,
to ensure that it is initialized.  That import used the predeclared
location, which caused various tests, notably Named_type::is_builtin,
to treat these imported names as builtin.  Start using a real location,
so that those tests do the right thing.

By the way, this implicit import is a partial cause of golang/go#19773.

Change-Id: I2137cefa679db519c63e17a3e94c1e5c4c5459a9
Reviewed-on: https://go-review.googlesource.com/c/gofrontend/+/194337
Reviewed-by: Than McIntosh <thanm@google.com>