gotest: force test package to be imported first
When compiling the x_test package, force the test package to be
imported first.  That ensures that we will see the types defined in
the test package before the types defined in the non-test version of
the package.  This matters if the types differ in some way, such as by
adding a new method.

This avoids a failure in internal/poll on Solaris, in which the test
package adds a method to a type (FD.EOFError).  I think it was Solaris-
specific because files are sorted in a different order by default.

The go tool handles this kind of thing correctly, by rebuilding
dependent packages.  This is just a hack sufficient to run the libgo
testsuite without using the go tool.

Fixes https://gcc.gnu.org/PR91712

Change-Id: I837ddd985a3a10e881c87cfe0dba4e6d369ab732
Reviewed-on: https://go-review.googlesource.com/c/gofrontend/+/194637
Reviewed-by: Than McIntosh <thanm@google.com>